Divide youth profile form into sections
