+++ src/domain/notFoundPage/NotFoundPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import PageContentWithHostingBox from '../../common/components/layout/PageContentWithHostingBox';
+
+function NotFoundPage() {
+  return <PageContentWithHostingBox>404 - not found</PageContentWithHostingBox>;
+}
+
+export default NotFoundPage;
