+++ src/domain/youthProfile/create/createYouthProfile.module.css
@@ -1,6 +1,5 @@
 @media (min-width: 768px){
     .form {
-        padding: 10px 50px;
         text-align: left;
     }
 }
